refactor: Исправить замечания по наименованию пакетов, удалению лишних зависимостей из pom.xml и инкриминации ID
